feat: support string enums
Relates to #29
