+++ src/emitter.test.ts
@@ -1,28 +1,5 @@
-import * as ts from 'typescript';
-
 import { toSource } from './index';
-
-export function getSourceFile(source: string, enableJsx = false): ts.SourceFile {
-  const options = {
-    ...ts.getDefaultCompilerOptions(),
-    experimentalDecorators: true,
-    jsx: ts.JsxEmit.Preserve,
-    noEmit: true,
-    noResolve: true,
-    preserveConstEnums: true,
-    removeComments: false,
-    target: ts.ScriptTarget.Latest
-  };
-  const host = {
-    ...ts.createCompilerHost(options, true),
-    getSourceFile(fileName: string, languageVersion: ts.ScriptTarget): ts.SourceFile {
-      return ts.createSourceFile(fileName, source, languageVersion, true);
-    }
-  };
-  const fileName = enableJsx ? 'source.tsx' : 'source.ts';
-  const program = ts.createProgram([fileName], options, host);
-  return program.getSourceFile(fileName);
-}
+import { getSourceFile } from './test-utils';
 
 describe('emit', () => {
   it('should accept a typescript as and reprint it', () => {
